requirements.txt
Django>=4.2,<5.0
djangorestframework
celery
redis
psycopg2-binary
gunicorn
python-dotenv
dj-database-url
